+++ spec/support/table_loader.rb
@@ -24,7 +24,13 @@ module TableLoader
       raise "Table config not found: #{path}" unless File.exist?(path)
 
       json = JSON.parse(File.read(path))
-      table = Exwiw::TableConfig.from_hash(json)
+      klass =
+        if adapter == :mongodb
+          Exwiw::MongodbCollectionConfig
+        else
+          Exwiw::TableConfig
+        end
+      table = klass.from(json)
       table_repository[adapter][table_name] = table
     end
   end